Closes EMB-816 misleading text for number of rows
