+++ frontend/src/metabase/common/utils/format-row-count.ts
@@ -1,0 +1,11 @@
+import { msgid, ngettext } from "ttag";
+
+import type { NumberFormatter } from "../hooks/use-number-formatter";
+
+export const formatRowCount = (
+  count: number,
+  formatNumber: NumberFormatter,
+) => {
+  const countString = formatNumber(count);
+  return ngettext(msgid`${countString} row`, `${countString} rows`, count);
+};
