+++ frontend/src/metabase/data-grid/components/DataGrid/DataGrid.tsx
@@ -52,6 +52,7 @@ export interface DataGridProps<TData>
     DataGridStylesProps {
   emptyState?: React.ReactNode;
   showRowsCount?: boolean;
+  rowsTruncated?: number;
   isColumnReorderingDisabled?: boolean;
   theme?: DataGridTheme;
   tableFooterExtraButtons?: React.ReactNode;
@@ -78,6 +79,7 @@ export const DataGrid = function DataGrid<TData>({
   onAddColumnClick,
   onWheel,
   tableFooterExtraButtons,
+  rowsTruncated,
 }: DataGridProps<TData>) {
   const {
     virtualColumns,
@@ -389,6 +391,7 @@ export const DataGrid = function DataGrid<TData>({
             table={table}
             enablePagination={enablePagination}
             showRowsCount={showRowsCount}
+            rowsTruncated={rowsTruncated}
             style={styles?.footer}
             className={classNames?.footer}
             tableFooterExtraButtons={tableFooterExtraButtons}
