+++ frontend/src/metabase/data-grid/components/Footer/Footer.tsx
@@ -3,6 +3,7 @@ import cx from "classnames";
 import { msgid, ngettext } from "ttag";
 
 import { useNumberFormatter } from "metabase/common/hooks/use-number-formatter";
+import { getRowCountMessage } from "metabase/common/utils/get-row-count-message";
 import { FOOTER_HEIGHT } from "metabase/data-grid/constants";
 import { PaginationFooter } from "metabase/visualizations/components/PaginationFooter/PaginationFooter";
 
@@ -12,6 +13,7 @@ export interface FooterProps<TData> {
   table: Table<TData>;
   enablePagination?: boolean;
   showRowsCount?: boolean;
+  rowsTruncated?: number;
   style?: React.CSSProperties;
   className?: string;
   tableFooterExtraButtons?: React.ReactNode;
@@ -24,6 +26,7 @@ export const Footer = <TData,>({
   className,
   style,
   tableFooterExtraButtons,
+  rowsTruncated,
 }: FooterProps<TData>) => {
   const formatNumber = useNumberFormatter();
   const wrapperAttributes = {
@@ -58,11 +61,19 @@ export const Footer = <TData,>({
       <div {...wrapperAttributes}>
         {tableFooterExtraButtons}
         <span className={S.rowsCount}>
-          {ngettext(
-            msgid`${formatNumber(total)} row`,
-            `${formatNumber(total)} rows`,
-            total,
-          )}
+          {rowsTruncated !== undefined
+            ? getRowCountMessage(
+                {
+                  data: { rows_truncated: rowsTruncated ?? 0 },
+                  row_count: total,
+                },
+                formatNumber,
+              )
+            : ngettext(
+                msgid`${formatNumber(total)} row`,
+                `${formatNumber(total)} rows`,
+                total,
+              )}
         </span>
       </div>
     );
