+++ frontend/src/metabase/visualizations/components/TableInteractive/TableInteractive.tsx
@@ -809,6 +809,7 @@ export const TableInteractiveInner = forwardRef(function TableInteractiveInner(
         {...tableProps}
         styles={dataGridStyles}
         showRowsCount={isDashboard}
+        rowsTruncated={data.rows_truncated}
         isColumnReorderingDisabled={isColumnReorderingDisabled}
         emptyState={emptyState}
         onBodyCellClick={handleBodyCellClick}
